--- OSMobile/osmobile/lib/app_widget/app_widget.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:osmobile/screens/login/login_screen.dart';
-
-class AppWidget extends StatelessWidget {
-  const AppWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-          primarySwatch: Colors.blue
-      ),
-      home: LoginScreen(),
-    );
-  }
-}
-
