+++ OSMobile/osmobile/lib/main.dart
@@ -1,6 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:osmobile/app_widget/app_widget.dart';
+
+import 'package:osmobile/pages/login/login_page.dart';
 
 void main() {
   runApp(const AppWidget());
+}
+
+class AppWidget extends StatelessWidget {
+  const AppWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      theme: ThemeData(primarySwatch: Colors.blue),
+      home: LoginScreen(),
+    );
+  }
 }
