+++ OSMobile/osmobile/lib/pages/Readme.txt
@@ -1,0 +1,1 @@
+Pasta para colocar as paginas do aplicativo.
